refactor: move stream body code from net/ to io/
