+++ io/httpbody.go
@@ -1,4 +1,4 @@
-package net
+package io
 
 import (
 	"bytes"
