+++ io/httpbody_test.go
@@ -1,4 +1,4 @@
-package net
+package io
 
 import (
 	"bytes"
