+++ proxy/proxy.go
@@ -33,10 +33,10 @@ import (
 	filterslog "github.com/zalando/skipper/filters/log"
 	ratelimitfilters "github.com/zalando/skipper/filters/ratelimit"
 	tracingfilter "github.com/zalando/skipper/filters/tracing"
+	skpio "github.com/zalando/skipper/io"
 	"github.com/zalando/skipper/loadbalancer"
 	"github.com/zalando/skipper/logging"
 	"github.com/zalando/skipper/metrics"
-	skpnet "github.com/zalando/skipper/net"
 	"github.com/zalando/skipper/proxy/fastcgi"
 	"github.com/zalando/skipper/ratelimit"
 	"github.com/zalando/skipper/rfc"
@@ -890,7 +890,7 @@ func (p *Proxy) makeBackendRequest(ctx *context, requestContext stdlibcontext.Co
 
 	ctx.proxySpan.LogKV("http_roundtrip", EndEvent)
 	if err != nil {
-		if errors.Is(err, skpnet.ErrBlocked) {
+		if errors.Is(err, skpio.ErrBlocked) {
 			p.tracing.setTag(ctx.proxySpan, BlockTag, true)
 			p.tracing.setTag(ctx.proxySpan, HTTPStatusCodeTag, uint16(http.StatusBadRequest))
 			return nil, &proxyError{err: err, code: http.StatusBadRequest}
